Improved system tag handling
If the RLogin username (provided by BBS) contiains a system tag,
then leave the username as is, otherwise prefix it with the
system tag from the config file.

Previously the system tag was being stripped from the RLogin
username (if present) and the tag from the config file was always
being prepended.

This accommodates those people using one instance of
doorparty-connector to forward connections for multiple boards.